Moves the admin User forms out of forms.py into admin.py
Lets keep admin functionality in admin.py
